LRAC-12122 fix browser and geolocation count type

+++ modules/dxp/apps/osb/osb-asah/osb-asah-common/src/main/java/com/liferay/osb/asah/common/repository/CustomBQSessionRepository.java
@@ -34,13 +34,13 @@ public interface CustomBQSessionRepository {
 
 	public List<BQSession> findAllById(Collection<String> sessionIds);
 
-	public Map<String, Integer> getSessionsCountGroupedByBrowserName(
+	public Map<String, BigDecimal> getSessionsCountGroupedByBrowserName(
 		Long channelId, TimeRange timeRange, ZoneId zoneId);
 
 	public List<Map<String, Object>> getSessionsCountGroupedByDeviceName(
 		Long channelId, TimeRange timeRange, ZoneId zoneId);
 
-	public Map<String, Integer> getSessionsCountGroupedByGeolocation(
+	public Map<String, BigDecimal> getSessionsCountGroupedByGeolocation(
 		Long channelId, TimeRange timeRange, ZoneId zoneId);
 
 	public List<SiteVisitorBehaviorMetric> getSiteVisitorBehaviorMetrics(
